moved the starlist file parser to backend
